backport of commit ab6fb6eb20a238fa4c849efaaa132b75e162561a
