+++ ui/lib/config-ui/addon/templates/messages/create.hbs
@@ -3,7 +3,7 @@
   SPDX-License-Identifier: BUSL-1.1
 ~}}
 
-<Messages::Page::CreateAndEditMessageForm
+<Messages::Page::CreateAndEdit
   @message={{this.model.message}}
   @messages={{this.model.messages}}
   @hasSomeActiveModals={{this.model.hasSomeActiveModals}}
